Dot Net 7 strangeness?

+++ src/Domain.LinnApps/Boms/ChangeRequestPhaseInWeeksView.cs
@@ -6,11 +6,11 @@
     {
         public string PhaseInWeek { get; set; }
 
-        public int? LinnWeekNumber { get; set; }
+        public decimal? LinnWeekNumber { get; set; }
 
         public DateTime? LinnEndDate { get; set; }
 
-        public int? DocumentNumber { get; set; }
+        public decimal? DocumentNumber { get; set; }
 
         public DateTime? DateAccepted { get; set; }
 
@@ -24,7 +24,7 @@
 
         public string DescriptionOfChange { get; set; }
 
-        public int? CountOfBomChanges { get; set; }
+        public decimal? CountOfBomChanges { get; set; }
 
         public string PhaseInType { get; set; }
 
@@ -32,8 +32,8 @@
 
         public string NewPartNumber { get; set; }
 
-        public int? OldPartStock { get; set; }
+        public decimal? OldPartStock { get; set; }
 
-        public int? NewPartStock { get; set; }
+        public decimal? NewPartStock { get; set; }
     }
 }
